Removed duplicate of translation declarations
